Fix decimal in remaining days and move badge to a small top banner above header

+++ core/resources/views/templates/basic/user/dashboard.blade.php
@@ -96,7 +96,7 @@
             </div>
             
             @php
-                $daysUsed = auth()->user()->created_at->diffInDays(now());
+                $daysUsed = (int) auth()->user()->created_at->diffInDays(now());
                 $isExpired = $daysUsed >= 30;
             @endphp
